Feat: 페이지와 좌석 선택 합체 & 좌석 선택 구현 - 내가 선택한 좌석 다른 선택자들에게 색 변경 - 내가 선택한 좌석 한번 더 클릭할 시 unsold 상태로 변경 - socket 위치 변경
