--- client/src/components/SeatSelectionArea/index.js
@@ -1,1 +1,0 @@
-export { default } from "./SeatSelectionArea.tsx"
